fix: import performance from perf_hooks and AggregateError from utils
